// src/app/main/sites/models/completeSite.ts
import { Site } from '../../../common/models/site';
import { SiteType } from './siteType';
import { Department } from '../../../common/models/department';
import { MealBonus } from './mealBonus';

export class CompleteSite extends Site {
  constructor() {
    super();
    this.mealBonus = null;
    this.specialties = null;
    this.isRestaurantVoucher = false;
    this.haveRestaurantVoucher = null;
    this.useScanner = null;
    this.departmentId = null;
    this.siteTypes = [];
    this.displayInTheWebSite = false;
    this.companyName = null;
    this.email = null;
    this.contact = null;
    this.webSite = null;
    this.department = null;
    this.name = null;
    this.city = null;
    this.zipCode = null;
    this.address = null;
    this.country = null;
    this.latitude = 0;
    this.longitude = 0;
    this.isIncludedInTraceability = false;
    this.fax = null;
    this.agreementNumber = null;
    this.veterinaryStamp = null;
    this.comment = null;
    this.active = true;
    this.activeString = null;
    this.phone = null;
    this.siteDescription = null;
  }
  mealBonus: MealBonus;
  specialties: string;
  isRestaurantVoucher: boolean;
  haveRestaurantVoucher: boolean | null;
  useScanner: boolean | null;
  departmentId: number | null;
  siteTypes: SiteType[];
  releaseDate: Date | string;
  entryDate: Date | string;
  displayInTheWebSite: boolean;
  companyName: string;
  email: string;
  contact: string;
  webSite: string;
  department: Department;
  name: string;
  city: string;
  zipCode: string;
  address: string;
  country: string;
  latitude: number;
  longitude: number;
  isIncludedInTraceability: boolean;
  fax: string;
  agreementNumber: string;
  veterinaryStamp: string;
  comment: string;
  active: boolean;
  activeString: string;
  phone: string;
  siteDescription: string;
  isActivityMealBag: boolean;
}